Fixed compile errors for ProductionCode trunk
git-svn-id: https://kwarqs-dev.bbn.com/svn/kwarqs/code@122 eb211378-2a44-443b-ae4a-d45323d45073

+++ ProductionCode/trunk/Framework/KwarqsWheelMotor.cpp
@@ -67,7 +67,7 @@ float KwarqsWheelMotor::GetSetSpeed()
 }
 
 /// Get the actual speed that the wheel is turning (via the encoder)
-float KwarqsWheelMotor::GetActualSpeed()
+double KwarqsWheelMotor::GetActualSpeed()
 {
 	// todo: implement me
 	return 0;
